Replace `rand_core`'s `std` feature with `alloc`
The `std` feature from `rand_core` activates the `alloc` feature and the
`getrandom` feature, which ends up including `getrandom` as a dependency.

For WebAssembly targets, it's only possible to obtain entropy if there's
an external function that provides that. Because of this, `getrandom`
only works if there are Javascript APIs enabled. If there aren't, it
fails to build.

The source of entropy isn't needed inside the library, therefore, the
`getrandom` feature does not need to be enabled. There is a use case in
a documentation test, which uses `OsRng`. However, it's still not
necessary to manually enable the `getrandom` feature in
`dev-dependencies`, because the `curve25519-dalek` dependency's
`default` feature enables `rand_core`'s `std` dependency.